+++ src/WatchCompass.Application/Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -4,6 +4,7 @@ using WatchCompass.Application.UseCases.MovieDetails;
 using WatchCompass.Application.UseCases.SearchMovies;
 using WatchCompass.Application.UseCases.Genres;
 using WatchCompass.Application.UseCases.SimilarMovies;
+using WatchCompass.Application.UseCases.TrendingMovies;
 
 namespace WatchCompass.Application.DependencyInjection;
 
@@ -16,6 +17,7 @@ public static class ApplicationServiceCollectionExtensions
         services.AddScoped<GetMovieDetailsUseCase>();
         services.AddScoped<GetGenresUseCase>();
         services.AddScoped<GetSimilarMoviesUseCase>();
+        services.AddScoped<GetTrendingMoviesUseCase>();
         return services;
     }
 }